requirements.txt
PyQt5>=5.15
PyOpenGL>=3.1
numpy>=1.23
trimesh>=4.0
# FBX Python SDK не публикуется в PyPI как пакет `fbx`.
# Поддержка FBX в приложении остается опциональной и отключается,
# если модуль не установлен вручную.
